rework send_to_recommender_reviewers_suggestions

+++ controllers/user_actions.py
@@ -194,39 +194,44 @@ def decline_review(): # no auth required
     quickDeclineKey = request.vars["key"]
 
     review = db.t_reviews[reviewId]
-    sendMessageForm = None
+
+    form = None
 
     if review is None:
         message = "Review '{}' not found".format(reviewId)
     elif review["review_state"] in ["Declined", "Declined manually", "Review completed", "Cancelled"]:
-        message = H4(T("You have already declined this invitation to review"), _class="decline-review-title")
-        user = db.auth_user[review.reviewer_id]
-        if user:
-            sendMessageForm = app_forms.getSendMessageForm(auth, db, user)
-
+        message = T("You have already declined this invitation to review")
     elif review.quick_decline_key != quickDeclineKey:
-        message = H4("Incorrect decline key: '{}'".format(quickDeclineKey), _class="decline-review-title")
+        message = "Incorrect decline key: '{}'".format(quickDeclineKey)
     else:
         review.review_state = "Declined"
         review.update_record()
 
         user = db.auth_user[review.reviewer_id]
-        if user:
-            sendMessageForm = app_forms.getSendMessageForm(auth, db, user)
+        if user and user.reset_password_key: # user (auto-created) did not register yet
+            db(db.auth_user.id == review.reviewer_id).delete()
 
-            if user.reset_password_key: # user (auto-created) did not register yet
-                db(db.auth_user.id == review.reviewer_id).delete()
-
-        message = H4(T("Thank you for taking the time to decline this invitation!"), _class="decline-review-title")
-
-    if sendMessageForm is not None:
-        if sendMessageForm.process(keepvalues=True).accepted:
-            suggRevText = sendMessageForm.vars.suggested_reviewers_text
-            emailing.send_to_recommender_reviewers_suggestions(session, auth, db, review.recommendation_id, suggRevText)
-            redirect(URL(c="default", f="index"))
+        message = T("Thank you for taking the time to decline this invitation!")
+        form = app_forms.getSendMessageForm(review.quick_decline_key)
 
     response.view = "default/info.html"
-    return dict(message=message, sendMessageForm=sendMessageForm)
+    return dict(
+        message=CENTER(
+            H4(message, _class="decline-review-title"),
+            form if form else DIV(_style="height: 20em;"),
+        )
+    )
+
+def send_suggested_reviewers():
+    text = request.post_vars.suggested_reviewers_text
+    review = db(db.t_reviews.quick_decline_key == request.post_vars.declineKey).select().last()
+
+    if not text or not review: redirect(URL(c="default", f="index"))
+
+    emailing.send_to_recommender_reviewers_suggestions(session, auth, db, review, text)
+
+    response.view = "default/info.html"
+    return dict(message=H4(T("Thank you for your suggestion!")), _class="decline-review-title")
 
 
 ######################################################################################################################################################################
